feat: split background jobs from shell toolset
